Refactor visualization library.
